swarm/storage: Rename resource to mru
