--- swarm/storage/resource/error.go
@@ -1,16 +1,0 @@
-package resource
-
-const (
-	ErrInit = iota
-	ErrNotFound
-	ErrIO
-	ErrUnauthorized
-	ErrInvalidValue
-	ErrDataOverflow
-	ErrNothingToReturn
-	ErrCorruptData
-	ErrInvalidSignature
-	ErrNotSynced
-	ErrPeriodDepth
-	ErrCnt
-)
